Optimization on tracking.
We have stopped tracking events that do not pertain to our invariant.

Clean up most of the inlines as we no longer need them.

This also makes setting up the transition easier as we don't need to keep
track of semi-colons during the Promela generation.

We still keep setRecentEvent as that one is multiple steps long. This keeps our
transitions readable.