Terminal state optimization. This is a small optimization compared to tracking.
Transitions and initial states now know whether the state they are transitioning
to is a terminal state.

This allows us to remove the extraneous terminal state tracking at the beginning
of every terminal state. This also leads us to removing the
"recentEvent.type = OTHEREVENT" as we have no event there.

That leads us to completely removing the artificial state we've been including
with every terminal state, leading to a more compact Promela model for Spin.